+++ CHANGELOG.md
@@ -5,3 +5,8 @@ First release.
 ## 0.1.0
 
 Add actual decoding and improve docs and example.
+
+## 0.1.1
+
+Fix handling of transparent pixels and add convenience function for scanning
+an encoded image in any format supported by the dart image library.